% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyODS.R
\name{.col2hex}
\alias{.col2hex}
\title{get hex representation of a color}
\usage{
.col2hex(colors)
}
\arguments{
\item{colors}{todo}
}
\value{
todo
}
\description{
get hex representation of a color
}
\examples{
prettyODS:::.col2hex(c("red","darkblue","#ff3300",NA))
}
