% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesAndSheets.R
\name{ODS_createStyle}
\alias{ODS_createStyle}
\title{Title}
\usage{
ODS_createStyle(
  font = NULL,
  size = NULL,
  color = NULL,
  bold = NULL,
  italic = NULL,
  underline = NULL,
  cellcolor = NULL,
  vAlign = NULL,
  hAlign = NULL,
  margin = NULL,
  wrap = NULL,
  rotate = NULL,
  digits = NULL
)
}
\arguments{
\item{font}{todo}

\item{size}{todo}

\item{color}{todo}

\item{bold}{todo}

\item{italic}{todo}

\item{underline}{todo}

\item{cellcolor}{todo}

\item{vAlign}{todo}

\item{hAlign}{todo}

\item{margin}{todo}

\item{wrap}{todo}

\item{rotate}{todo}

\item{digits}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
ODS_createStyle(font="Arial",bold=TRUE)
}
