% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyODS.R
\name{ODS_getStyle}
\alias{ODS_getStyle}
\title{Title}
\usage{
ODS_getStyle(sheet, row, col)
}
\arguments{
\item{sheet}{todo}

\item{row}{todo}

\item{col}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
1+1
}
