% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyODS.R
\name{ODS_mergeStyles}
\alias{ODS_mergeStyles}
\title{Title}
\usage{
ODS_mergeStyles(style1, style2)
}
\arguments{
\item{style1}{todo}

\item{style2}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
1+1
}
