% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{ODS_writeCell}
\alias{ODS_writeCell}
\title{Title}
\usage{
ODS_writeCell(sheet, data, row, col, style)
}
\arguments{
\item{sheet}{todo}

\item{data}{todo}

\item{row}{todo}

\item{col}{todo}

\item{style}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
1+1
}
