% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{.ODS_writeArray}
\alias{.ODS_writeArray}
\title{Title}
\usage{
.ODS_writeArray(sheet, array, startRow, startCol, style, skipNA = FALSE)
}
\arguments{
\item{sheet}{todo}

\item{array}{todo todo}

\item{startRow}{todo}

\item{startCol}{todo}

\item{style}{todo}

\item{skipNA}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
1+1
}
