% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{ODS_mergeCells}
\alias{ODS_mergeCells}
\title{Title}
\usage{
ODS_mergeCells(sheet, rows, cols, mergeCols = TRUE, mergeRows = TRUE)
}
\arguments{
\item{sheet}{todo}

\item{rows}{todo}

\item{cols}{todo}

\item{mergeCols}{todo}

\item{mergeRows}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
1+1
}
