% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyODS.R
\name{ODS_setColWidths}
\alias{ODS_setColWidths}
\title{Title}
\usage{
ODS_setColWidths(sheet, cols, width = 1.7, unit = "cm")
}
\arguments{
\item{sheet}{todo}

\item{cols}{todo}

\item{width}{todo}

\item{unit}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
1+1
}
