% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyODS.R
\name{ODS_setRowHeights}
\alias{ODS_setRowHeights}
\title{Title}
\usage{
ODS_setRowHeights(sheet, rows, height = 15, unit = "pt")
}
\arguments{
\item{sheet}{todo}

\item{rows}{todo}

\item{height}{todo}

\item{unit}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
1+1
}
