% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{ODS_writeData}
\alias{ODS_writeData}
\title{Title}
\usage{
ODS_writeData(
  sheet,
  data,
  startRow = 1,
  startCol = 1,
  style,
  useColnames = TRUE,
  styleColnames
)
}
\arguments{
\item{sheet}{todo}

\item{data}{todo todo}

\item{startRow}{todo}

\item{startCol}{todo}

\item{style}{todo}

\item{useColnames}{todo}

\item{styleColnames}{todo}
}
\value{
todo
}
\description{
Title
}
\examples{
1+1
}
